+++ src/ffi/mod.rs
@@ -12,3 +12,4 @@ pub(crate) mod network_sets;
 pub(crate) mod preferences;
 pub(crate) mod schema;
 pub(crate) mod system_configuration;
+pub(crate) mod async_api;